+++ cli/commands/print/chains.ts
@@ -4,7 +4,7 @@ import { Command } from "commander";
 export const chainCmd = new Command("chains").description("List all supported chain slugs").action(() => {
   console.log("🌐 Mainnets:");
   console.log(
-    sablier.chains
+    sablier.evm.chains
       .getMainnets()
       .map((c) => `• ${c.slug}`)
       .join("\n"),
@@ -12,7 +12,7 @@ export const chainCmd = new Command("chains").description("List all supported ch
 
   console.log("\n🧪 Testnets:");
   console.log(
-    sablier.chains
+    sablier.evm.chains
       .getTestnets()
       .map((c) => `• ${c.slug}`)
       .join("\n"),