+++ src/evm/releases/lockup/v1.2/index.ts
@@ -1,0 +1,22 @@
+import { Protocol } from "@src/evm/enums";
+import { sortDeployments } from "@src/evm/releases/helpers";
+import { resolvers } from "@src/evm/releases/resolvers";
+import type { Sablier } from "@src/types";
+import { abi } from "./abi";
+import aliases from "./aliases";
+import { mainnets, testnets } from "./deployments";
+import manifest from "./manifest";
+
+const sortedMainnets = sortDeployments(mainnets);
+const sortedTestnets = sortDeployments(testnets);
+const deployments: Sablier.EVM.Deployment.LockupV1[] = [...sortedMainnets, ...sortedTestnets];
+
+export const release = resolvers.release.lockupV1({
+  abi,
+  aliases,
+  deployments,
+  isLatest: false,
+  manifest,
+  protocol: Protocol.Lockup,
+  version: "v1.2",
+});