+++ src/sablier.ts
@@ -6,7 +6,15 @@
  * ```typescript
  * import { sablier } from "sablier";
  *
- * const lockupContract = sablier.contracts.get({
+ * // EVM
+ * let lockupContract = sablier.evm.contracts.get({
+ *   chainId: mainnet.id,
+ *   contractName: "SablierLockup",
+ *   release: releases.lockup["v2.0"],
+ * });
+ *
+ * // Can also be accessed like this:
+ * lockupContract = sablier.contracts.get({
  *   chainId: mainnet.id,
  *   contractName: "SablierLockup",
  *   release: releases.lockup["v2.0"],
@@ -16,28 +24,36 @@
  */
 import type { Sablier } from "@src/types";
 import _ from "lodash";
-import { chainsQueries } from "./chains/queries";
-import { contractsQueries } from "./contracts/queries";
-import { releasesQueries } from "./releases/queries";
+import { chainsQueries as evmChainsQueries } from "./evm/chains/queries";
+import { contractsQueries as evmContractsQueries } from "./evm/contracts/queries";
+import { releasesQueries as evmReleasesQueries } from "./evm/releases/queries";
 
-const deploymentsQueries = {
+/**
+ * Has to be defined here to avoid circular dependencies.
+ */
+const evmDeploymentsQueries = {
   /**
    * Get many deployments.
    * - default            ⇒ all across all releases
    * - release            ⇒ that release's deployments
    */
-  get: (opts: { chainId: number; release: Sablier.Release }): Sablier.Deployment | undefined => {
+  get: (opts: { chainId: number; release: Sablier.EVM.Release }): Sablier.EVM.Deployment | undefined => {
     const { release, chainId } = opts || {};
     return _.find(release.deployments, { chainId });
   },
-  getAll: (): Sablier.Deployment[] => {
-    return _.flatMap(releasesQueries.getAll(), (r) => r.deployments);
+  getAll: (): Sablier.EVM.Deployment[] => {
+    return _.flatMap(evmReleasesQueries.getAll(), (r) => r.deployments);
   },
 };
 
+const evm = {
+  chains: evmChainsQueries,
+  contracts: evmContractsQueries,
+  deployments: evmDeploymentsQueries,
+  releases: evmReleasesQueries,
+};
+
 export const sablier = {
-  chains: chainsQueries,
-  contracts: contractsQueries,
-  deployments: deploymentsQueries,
-  releases: releasesQueries,
+  ...evm, // re-exporting for backward compatibility
+  evm,
 };