+++ tests/contracts/checksums.test.ts
@@ -6,9 +6,9 @@ import { describe, expect, it } from "vitest";
  * @note Use https://ethsum.netlify.app to fix failing tests
  */
 describe("Address checksums", () => {
-  for (const release of sablier.releases.getAll()) {
+  for (const release of sablier.evm.releases.getAll()) {
     describe(`${release.protocol} ${release.version}`, () => {
-      const contracts = sablier.contracts.getAll({ release })!;
+      const contracts = sablier.evm.contracts.getAll({ release })!;
 
       for (const contract of contracts) {
         it(`${contract.name} should have a checksummed address`, () => {